Added Endpoints for Eigenschaften

+++ DSA5.Application/Editor/Rassen/Handlers/DeleteSpeziesHandler.cs
@@ -1,11 +1,11 @@
 ﻿using DSA5.Application.Common.Persistence;
-using DSA5.Application.Editor.BaseEntities.Rassen.Commands;
+using DSA5.Application.Editor.Rassen.Commands;
 using DSA5.Entities.Welt;
 using MediatR;
 
-namespace DSA5.Application.Editor.BaseEntities.Rassen.CommandHandlers;
+namespace DSA5.Application.Editor.Rassen.Handlers;
 
-public sealed class DeleteSpeziesHandler : IRequestHandler<DeleteSpezies>
+public sealed class DeleteSpeziesHandler : IRequestHandler<DeleteSpeziesRequest>
 {
     private readonly IRepository<Spezies> _repository;
 
@@ -14,9 +14,10 @@ public sealed class DeleteSpeziesHandler : IRequestHandler<DeleteSpezies>
         _repository = repository;
     }
 
-    public async Task Handle(DeleteSpezies request, CancellationToken cancellationToken)
+    public async Task Handle(DeleteSpeziesRequest request, CancellationToken cancellationToken)
     {
         var spezies = await _repository.GetByIdAsync(request.Id, cancellationToken);
+        if (spezies is null) return;
         await _repository.DeleteAsync(spezies, cancellationToken);
     }
 }
